requirements.txt
python-telegram-bot==20.6
openai==1.14.3
apscheduler==3.10.4
nest_asyncio==1.6.0
